app/src/main/java/com/ashehata/testme/room/UserDao.kt
package com.ashehata.testme.room

import androidx.room.Dao
import androidx.room.Delete
import androidx.room.Insert
import androidx.room.Query
import kotlinx.coroutines.flow.Flow

typealias Users = List<User>

@Dao
interface UserDao {
    @Query("SELECT * FROM user")
    fun getAll(): Flow<List<User>>

    @Query("SELECT * FROM user WHERE uid IN (:userIds)")
    fun loadAllByIds(userIds: IntArray): Flow<Users>

    @Query("SELECT * FROM user WHERE first_name LIKE :first AND " +
            "last_name LIKE :last LIMIT 1")
    fun findByName(first: String, last: String): Flow<User>

    @Insert
    suspend fun insert(user: User)

    @Insert
    suspend fun insertAll(users: Users)

    @Delete
    suspend fun delete(user: User)
}